Refactor authentication and registration URLs: remove AddFaceEncoding endpoint from authentication and add it to registration; update imports accordingly.
