+++ api/urls/authentication.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from api.views import StudentLoginEndPoint, TeacherLoginEndPoint, AdminSignInEndPoint, LogoutEndPoint, RegenerateAccessToken, AddFaceEncodingEndPoint, LoginFaceEndPoint
+from api.views import StudentLoginEndPoint, TeacherLoginEndPoint, AdminSignInEndPoint, LogoutEndPoint, RegenerateAccessToken, LoginFaceEndPoint
 
 urlpatterns = [
     path('student/login/', StudentLoginEndPoint.as_view()),
@@ -7,6 +7,5 @@ urlpatterns = [
     path('adminendpoint/login/', AdminSignInEndPoint.as_view()),
     path('logout/', LogoutEndPoint.as_view()),
     path('token/refresh/', RegenerateAccessToken.as_view()),
-    path('add_pfp_and_face_encoding/', AddFaceEncodingEndPoint.as_view()),
     path('student/login_face/', LoginFaceEndPoint.as_view()),
 ]
